refactor: use daily statistic generator service
